Fully evaluate `heroku_apps` before use
`get_apps_for_teams` is a generator, meaning the second iteration to define `known_heroku_apps` would be empty as the generator was already exhausted.

This lead to the script progressively removing and recreating DNS records with each execution, which could make it look like the site is down, without errors.

It's always DNS!!!